LayoutTests/fast/js/kde/scope-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f(2) is 22
PASS successfullyParsed is true

TEST COMPLETE
